refactor: split widget extension rendering responsibilities
